Narrow gr.render HTML remount fix
